[Refactor][Core/Dashboard] Merge ActorHead into NodeHead
