+++ python/ray/dashboard/modules/node/actor_consts.py
@@ -1,5 +1,4 @@
 import ray
 
-ACTOR_CHANNEL = "ACTOR"
 NIL_NODE_ID = ray.NodeID.nil().hex()
 RETRY_GET_ALL_ACTOR_INFO_INTERVAL_SECONDS = 1
